Rename QueriesSaveButtons to SaveButton and use also for Scripts
