+++ src/scripts/modules/transformations/react/pages/transformation-detail/Scripts.jsx
@@ -1,8 +1,7 @@
-import React, {PropTypes} from 'react';
-import { FormGroup, FormControl, ControlLabel } from 'react-bootstrap';
+import React, { PropTypes } from 'react';
 import Edit from './ScriptsEdit';
 import Clipboard from '../../../../../react/common/Clipboard';
-import SaveButtons from '../../../../../react/common/SaveButtons';
+import SaveButton from '../../components/SaveButton';
 
 /* global require */
 require('codemirror/mode/r/r');
@@ -26,10 +25,10 @@ export default React.createClass({
   render() {
     return (
       <div>
-        <h2 style={{lineHeight: '32px'}}>
+        <h2 style={{ lineHeight: '32px' }}>
           Scripts
           <small>
-            <Clipboard text={this.props.scripts}/>
+            <Clipboard text={this.props.scripts} />
           </small>
           {this.renderButtons()}
         </h2>
@@ -41,27 +40,14 @@ export default React.createClass({
   renderButtons() {
     return (
       <span className="pull-right">
-        <SaveButtons
+        <SaveButton
           isSaving={this.props.isSaving}
-          disabled={!this.props.isEditingValid}
           isChanged={this.props.isChanged}
           onSave={this.props.onEditSubmit}
           onReset={this.props.onEditCancel}
-          showModal={true}
-          modalTitle="Save new scripts"
-          modalBody={
-            <FormGroup>
-              <ControlLabel>Update description</ControlLabel>
-              <FormControl
-                componentClass="textarea"
-                rows={3}
-                autoFocus
-                value={this.props.changeDescription}
-                onChange={e => this.props.onDescriptionChange(e.target.value)}
-                disabled={this.props.isSaving}
-              />
-            </FormGroup>
-          }
+          disabled={!this.props.isEditingValid}
+          onDescriptionChange={this.props.onDescriptionChange}
+          changeDescription={this.props.changeDescription}
         />
       </span>
     );
